put everything inside if clause

+++ app/models/spina/page.rb
@@ -115,9 +115,11 @@ module Spina
     end
 
     def rewrite_rule
-      RewriteRule.where(old_path: materialized_path).delete_all
-      RewriteRule.where(new_path: old_path).update(new_path: materialized_path)
-      RewriteRule.where(old_path: old_path).first_or_create.update(new_path: materialized_path) if old_path != materialized_path
+      if old_path != materialized_path
+        RewriteRule.where(old_path: materialized_path).delete_all
+        RewriteRule.where(new_path: old_path).update(new_path: materialized_path)
+        RewriteRule.where(old_path: old_path).first_or_create.update(new_path: materialized_path)
+      end
     end
 
     def localized_materialized_path
